# Speed up calculation of libzypp product renames