--- a/package/yast2-packager.changes
+++ b/package/yast2-packager.changes
@@ -1,4 +1,10 @@
 -------------------------------------------------------------------
+Thu Dec  5 18:20:30 UTC 2019 - Imobach Gonzalez Sosa <igonzalezsosa@suse.com>
+
+- Speed up product renames calculation (bsc#1157926).
+- 4.1.50
+
+-------------------------------------------------------------------
 Thu Aug 29 11:11:05 UTC 2019 - José Iván López González <jlopez@suse.com>
 
 - Avoid error when generating some warnings (bsc#1148536).