--- a/package/yast2-packager.spec
+++ b/package/yast2-packager.spec
@@ -17,7 +17,7 @@
 
 
 Name:           yast2-packager
-Version:        4.1.49
+Version:        4.1.50
 Release:        0
 
 BuildRoot:      %{_tmppath}/%{name}-%{version}-build