--- a/test/addon_product_test.rb
+++ b/test/addon_product_test.rb
@@ -31,6 +31,16 @@ describe Yast::AddOnProduct do
     end
 
     context "when according to libzypp a product is renamed" do
+      before do
+        subject.main
+        allow(Y2Packager::Repository).to receive(:all).and_return([repo0])
+      end
+
+      let(:repo0) do
+        instance_double(
+          Y2Packager::Repository, repo_id: 0, url: URI("dvd:///sr0"), product_dir: "/p0"
+        )
+      end
       let(:deps) do
         [
           { "obsoletes" => "product:old_product1" },
@@ -62,6 +72,37 @@ describe Yast::AddOnProduct do
         expect(subject.renamed?("old_product1", new_product["name"])).to eq(true)
         expect(subject.renamed?("old_product2", new_product["name"])).to eq(true)
         expect(subject.renamed?("old_name", new_product["name"])).to eq(true)
+      end
+
+      context "when renames information has been already loaded" do
+        before do
+          subject.renamed?("old_name", new_product["name"])
+        end
+
+        it "does not ask libzypp again" do
+          expect(Yast::Pkg).to_not receive(:ResolvableDependencies)
+          subject.renamed?("old_name", new_product["name"])
+        end
+
+        context "but renames information is obsolete" do
+          let(:repo1) do
+            instance_double(
+              Y2Packager::Repository, repo_id: 1, url: URI("dvd:///sr0"), product_dir: "/p1"
+            )
+          end
+
+          before do
+            subject.renamed?("old_name", new_product["name"])
+            allow(Y2Packager::Repository).to receive(:all).and_return([repo0, repo1])
+          end
+
+          it "asks libzypp again" do
+            expect(Yast::Pkg).to receive(:ResolvableDependencies)
+              .with(new_product["product_package"], :package, "")
+              .and_return([installed_product_package, new_product_package])
+            subject.renamed?("old_name", new_product["name"])
+          end
+        end
       end
     end
   end