--- a/test/addon_product_test.rb
+++ b/test/addon_product_test.rb
@@ -2,6 +2,7 @@
 
 require_relative "./test_helper"
 require_relative "product_factory"
+require "uri"
 
 Yast.import "AddOnProduct"
 
@@ -84,7 +85,7 @@ describe Yast::AddOnProduct do
           subject.renamed?("old_name", new_product["name"])
         end
 
-        context "but renames information is obsolete" do
+        context "but a new repo has been added" do
           let(:repo1) do
             instance_double(
               Y2Packager::Repository, repo_id: 1, url: URI("dvd:///sr0"), product_dir: "/p1"
@@ -92,14 +93,44 @@ describe Yast::AddOnProduct do
           end
 
           before do
-            subject.renamed?("old_name", new_product["name"])
             allow(Y2Packager::Repository).to receive(:all).and_return([repo0, repo1])
           end
 
           it "asks libzypp again" do
-            expect(Yast::Pkg).to receive(:ResolvableDependencies)
-              .with(new_product["product_package"], :package, "")
-              .and_return([installed_product_package, new_product_package])
+            expect(Yast::Pkg).to receive(:ResolvableDependencies).and_return([])
+            subject.renamed?("old_name", new_product["name"])
+          end
+        end
+
+        context "but the repo_id for a given repo has changed" do
+          before do
+            allow(repo0).to receive(:repo_id).and_return(1)
+          end
+
+          it "asks libzypp again" do
+            expect(Yast::Pkg).to receive(:ResolvableDependencies).and_return([])
+            subject.renamed?("old_name", new_product["name"])
+          end
+        end
+
+        context "but the url for a given repo has changed" do
+          before do
+            allow(repo0).to receive(:url).and_return(URI("dvd:///sr2"))
+          end
+
+          it "asks libzypp again" do
+            expect(Yast::Pkg).to receive(:ResolvableDependencies).and_return([])
+            subject.renamed?("old_name", new_product["name"])
+          end
+        end
+
+        context "but the product_dir for a given repo has changed" do
+          before do
+            allow(repo0).to receive(:product_dir).and_return("/another")
+          end
+
+          it "asks libzypp again" do
+            expect(Yast::Pkg).to receive(:ResolvableDependencies).and_return([])
             subject.renamed?("old_name", new_product["name"])
           end
         end