This PR solves [bsc#1157926](https://bugzilla.suse.com/show_bug.cgi?id=1157926). The real problem is that it takes too much time to calculate the product renames from `libzypp`. So this PRs adds some code to cache these renames.

On the other hand,  it avoids trying to find the product rename when the `product_package` property is missing, avoiding a lot of failing calls to `Pkg.ResolvableDependencies`.

In a test I did in a VM, the patch reduced the time to show the analysis result from 45 seconds to just 9 seconds. We can even improve this time by moving the logic to invalidate the cached data outside of the `AddOnProduct` module, avoiding asking libzypp for changes. However, that approach looks wrong to me.

Trello: https://trello.com/c/ZTskvrjr/

![analyzing-your-system](https://user-images.githubusercontent.com/15836/70262007-6c1db280-178b-11ea-9857-25dd5d949fa5.png)

## Speed comparison

<details>
<summary>Migration proposal (from 45s to 9s)</summary>

Old code: ~45 seconds
![load-migration-products](https://user-images.githubusercontent.com/15836/70532580-41e34080-1b4f-11ea-8b2a-c3e52081f6fe.gif)

New code: ~9 seconds
![faster-load-migration-products](https://user-images.githubusercontent.com/15836/70532597-4d366c00-1b4f-11ea-89af-d0f5c1d5b2b5.gif)
</details>

<details>
<summary>Migration target selection (from 12s to 4s)</summary>

Old code
![load-migration-products](https://user-images.githubusercontent.com/15836/70527673-5f5edd00-1b44-11ea-9e8a-bd14902cc12c.gif)

New code
![faster-load-migration-products](https://user-images.githubusercontent.com/15836/70527701-6e458f80-1b44-11ea-8da4-5ca9d7d7ecf1.gif)
</details>