@imobachgs 2019-12-10
> Maybe changing

[..]

> could speed it up a bit more as this would not fetch the unused attributes (like RPM size, path on the medium, package license, etc...) from libzypp.

Yes, perhaps it could help. Unfortunately, the new resolvables API is not available in SLE 15 SP1.